+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 - Navigation through listings using label names and CSECT offsets
 - Support analyzing files opened directly by Zowe Explorer
+- Progress indicator
 
 #### Fixed
 - Symbols in listings with ASA characters are incorrectly isolated